Updated Bot.js, Created Help.js
Updated bot.js to have emergency commented code into it just in case I need to reinitialize the application commands.
Created Help.js which contains the code for the help command which goes through all of the available application commands and displays them onto an embed and limits to 5 commands per page, once there is 5 commands on one page it will create an new page and enable the button to be pressed so you can then view the rest of the commands.