Create punishment-lookup.js
Using mongodb when an user has an punishment given to them an random string of letters is generated and called the punishment ID, this ID is sent to the responsible moderator of the punishment and the target user dms. You can then use this command along side the punishment ID to lookup the details of the punishment.