Updated Help.js Created TictacToe.js
I have updated to Help.js so it actually works when it there is more than 5 commands and someone needs to click on the next button to switch to the next page to view the other commands.
Created TicTacToe, using the discord-tictactoe library I have created an tictactoe command that allows you to start an tictactoe game with you and the bot or you and someone that you mention in the target-user option.