Updated Ban.js, Created Kick.js
Updated Ban.js to fix an grammar error in the option description of the reason option.
Created Kick.js uses the same logic as the ban command with the mongo db using punishment ids etc except this one just kicks the user and changes the punishment type to be an kick instead of an ban.