Created level.js Created Level.js Created giveUserXp.js
Create level.js the main command of checking your level or someone elses level. Uses canvacord to create an image of the level. Created Level.js this is just the model file for the level document.  Created giveUserXp.js this just uses an random num generated of 5xp to 15xp when an new message is typed and puts that user on an cooldown, adds the selected xp to the database.