Create automod-enable.js Create autoMod.js Create AutoMod.js Edited tictactoe.js
Create automod-enable.js if the automod feature is disabled it sets the automodEnabled value in my database to be set to true so the automod feature will work for that guild.
Create autoMod.js this is the event file that checks through each message that is sent and checks to see if any of the messages apply to any of the regexs which are: youtubeLinks discordInviteLinks twitchLinks and racialSlurs.
Create AutoMod.js this is the model file for the AutoMod document in the mongodb database.
Updated tictactoe.js had to fix one small issue with my capitals in an argument for an function.