Create automod-disable updated automod-enable
Created automod-disable same logic as the enable command apart from it just sets the automodEnabled entry in the database to be false instead of true.
Updated automod-enable.js edited the messages that are sent when its enabled to show what command to run if they want to disable the automod.