Create TicketInfo.js Create ticket.js Create ticket-configure.js
Created the TicketInfo model for the ticket feature.
Created ticket-configure.js the command that is used to setup the tickets feature only argument is the roleId that would be allowed access to the tickets that are created.
Create ticket.js the command that is ran by the user that creates the ticket channel.