Create warn.js
Created an warn command that warns the specified user and yet again stores it in the database, but this time it checks to see if the user has 3 warns and if they do then it kicks them from the server for reaching 3 warns in the server. If they don't then it just warns them from the server.